Change float64_t arg-type for set_throttle_threshold to float32_t
The C++ method signature of set_throttle_threshold/set_trottle_recovery_delay,
which uses nvbench::float32_t

+++ python/src/py_nvbench.cpp
@@ -376,7 +376,7 @@ PYBIND11_MODULE(_nvbench, m)
     py::arg("duration_seconds"));
   py_benchmark_cls.def(
     "set_throttle_threshold",
-    [](nvbench::benchmark_base &self, nvbench::float64_t threshold) {
+    [](nvbench::benchmark_base &self, nvbench::float32_t threshold) {
       self.set_throttle_threshold(threshold);
       return std::ref(self);
     },
@@ -384,7 +384,7 @@ PYBIND11_MODULE(_nvbench, m)
     py::arg("threshold"));
   py_benchmark_cls.def(
     "set_throttle_recovery_delay",
-    [](nvbench::benchmark_base &self, nvbench::float64_t delay) {
+    [](nvbench::benchmark_base &self, nvbench::float32_t delay) {
       self.set_throttle_recovery_delay(delay);
       return std::ref(self);
     },
